Added ranged_integer overloads for OutMessage member functions.

+++ source/clients/pokemon_online/outmessage.cpp
@@ -84,7 +84,7 @@ void OutMessage::write_pokemon (Pokemon const & pokemon) {
 	bool shiny = false;
 	write_byte (shiny);
 	write_byte(pokemon.happiness());
-	write_byte(static_cast<uint8_t>(get_level(pokemon)()));
+	write_byte(get_level(pokemon)());
 	unsigned number_of_moves = 0;
 	pokemon.move.for_each_regular_move([&](Move const & move) {
 		++number_of_moves;
